Fixed some mistakes in shaders. TODO: Skybox subroutine won't work on NVIDIA..

+++ engine/shaders/basiclightning.frag
@@ -80,7 +80,7 @@ uniform vec3 gEyeWorldPos;
 uniform bool gHasTangents;
 uniform bool gShadowsEnabled;
 
-float calcShadowFactor(in vec4 lightSpacePos, in sampler2D shadowMap, unsigned int size)
+float calcShadowFactor(in vec4 lightSpacePos, in sampler2D shadowMap, float size)
 {
     if(!gShadowsEnabled)
     {
